requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cffi==1.15.1
charset-normalizer==3.1.0
discord.py==2.3.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.0
yarl==1.9.2
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
discord.py==2.3.1
Flask==2.3.2
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.0
Werkzeug==2.3.6
yarl==1.9.2


